+++ iOS-sdk/SessionM_MMC_iOS_v1.15.5_Release/SessionM-SDK/API/SessionM.h
@@ -1164,7 +1164,7 @@ typedef enum SMAchievementDismissType {
  @property tierPointBalance
  @abstract Returns user's current tier points balance.
  */
-@property(nonatomic, readonly) NSUInteger tierPointBalance;
+@property(nonatomic, readonly) unsigned long tierPointBalance;
 /*!
  @property unclaimedAchievementCount
  @abstract Returns user's unclaimed achievement count.
